Create Businesser
This project was done as an extension of my work as a financial data analyst. A number of simple insights from this experience have been included in this project. A number of assumptions were made on how much weight should be applied to each category as well as how each category should be graded. Most of these assumptions are subjective depending on any number of variables.